Secure login changes
Add a secure login requirement for moderators. Secure login requirement means 2FA cannot be removed (but password overall can) and there is a redirect on login if secure requirements are not met.

Additionally, allow users with 2FA set up to regenerate the code. This is needed for those with secure logins as they can't remove and readd.